// contents/__generated__/article-nav.ts
// This file is auto-generated
import type { Globals } from "@/types/globals";

export const articleNav: Globals.Data.ArticleNavProps[] = [
  {
    "key": "koloniyalizm",
    "title": "Koloniyalizm",
    "path": "/konular/koloniyalizm",
    "subtopics": [
      {
        "key": "afrikada-kolonicilik",
        "title": "Afrika'da Kolonicilik",
        "path": "/konular/koloniyalizm/afrikada-kolonicilik",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/afrikada-kolonicilik/example"
          }
        ]
      },
      {
        "key": "amerikada-kolonicilik",
        "title": "Amerika'da Kolonicilik",
        "path": "/konular/koloniyalizm/amerikada-kolonicilik",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/amerikada-kolonicilik/example"
          }
        ]
      },
      {
        "key": "asyada-kolonicilik",
        "title": "Asya'da Kolonicilik",
        "path": "/konular/koloniyalizm/asyada-kolonicilik",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/asyada-kolonicilik/example"
          }
        ]
      },
      {
        "key": "dekolonizasyon",
        "title": "Dekolonizasyon",
        "path": "/konular/koloniyalizm/dekolonizasyon",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/dekolonizasyon/example"
          }
        ]
      },
      {
        "key": "direnisler-ve-isyanlar",
        "title": "Direnişler ve İsyanlar",
        "path": "/konular/koloniyalizm/direnisler-ve-isyanlar",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/direnisler-ve-isyanlar/example"
          }
        ]
      },
      {
        "key": "imparatorluklar",
        "title": "İmparatorluklar",
        "path": "/konular/koloniyalizm/imparatorluklar",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/imparatorluklar/example"
          }
        ]
      },
      {
        "key": "kesif-cagi",
        "title": "Keşif Çağı",
        "path": "/konular/koloniyalizm/kesif-cagi",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/kesif-cagi/example"
          }
        ]
      },
      {
        "key": "pasifikte-kolonicilik",
        "title": "Pasifik'te Kolonicilik",
        "path": "/konular/koloniyalizm/pasifikte-kolonicilik",
        "articles": [
          {
            "key": "example",
            "title": "Example",
            "path": "/konular/koloniyalizm/pasifikte-kolonicilik/example"
          }
        ]
      }
    ]
  }
];